Downloads/system-builder-hub/backend/requirements-optional.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Environment and Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Database and ORM
SQLAlchemy==2.0.21
alembic==1.12.0

# CLI and Operations
click==8.1.7

# System Monitoring and Metrics
psutil==5.9.5

# Security and Cryptography
cryptography==41.0.4

# HTTP and Networking
requests==2.31.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# AI and Machine Learning
openai==0.28.1
anthropic==0.7.7

# File Processing
Pillow==10.0.1

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0

# Optional: For advanced features
# playwright==1.39.0  # For screenshot functionality
# docker==6.1.3       # For containerized previews

# Audio Processing (Optional - platform-specific)
# pyaudio==0.2.11     # For voice input processing (requires system audio libraries)
